Fixed `heredoc`
`open_error` still exits in that one case
Have to remove `t_cmdlist` and change `content` to `main_arr`
