--- src/frees2.c
@@ -1,26 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   frees2.c                                           :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/28 17:51:12 by mortins-          #+#    #+#             */
-/*   Updated: 2023/10/30 18:37:41 by mortins-         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../inc/minishell.h"
-
-void	free_list_and_data(t_list **list)
-{
-	t_list	*tmp;
-
-	while (*list)
-	{
-		tmp = *list;
-		*list = (*list)->next;
-		free(tmp->data);
-		free(tmp);
-	}
-}
