+++ src/run/redirections.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   redirections.c                                     :+:      :+:    :+:   */
+/*   new_redirections.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/20 17:36:41 by mortins-          #+#    #+#             */
-/*   Updated: 2023/10/30 16:29:11 by mortins-         ###   ########.fr       */
+/*   Created: 2023/11/01 15:54:54 by mortins-          #+#    #+#             */
+/*   Updated: 2023/11/01 16:44:06 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,63 +20,55 @@ void	reset_fds(t_minishell *ms)
 	close(ms->fdin_buf);
 }
 
-void	redirect_out(t_list *out, int append)
+void	redirect_in(char *file, int heredoc)
 {
 	int	fd;
 
-	while (out)
+	fd = open(file, O_RDONLY);
+	if (fd < 0)
+		open_error(file);
+	else
 	{
-		if (append == 0)
-			fd = open(out->data, O_CREAT | O_RDWR | O_TRUNC, 0664);
-		else
-			fd = open(out->data, O_CREAT | O_RDWR | O_APPEND, 0664);
-		if (fd < 0)
-			open_error(out->data);
-		else
-		{
-			dup2(fd, STDOUT_FILENO);
-			close(fd);
-		}
-		out = out->next;
+		dup2(fd, STDIN_FILENO);
+		close(fd);
+		if (heredoc)
+			unlink(file);
 	}
 }
 
-void	redirect_in(t_list *in)
+void	redirect_out(char *file, int append)
 {
 	int	fd;
 
-	while (in)
+	if (!append)
+		fd = open(file, O_CREAT | O_RDWR | O_TRUNC, 0664);
+	else
+		fd = open(file, O_CREAT | O_RDWR | O_APPEND, 0664);
+	if (fd < 0)
+		open_error(file);
+	else
 	{
-		fd = open(in->data, O_RDONLY);
-		if (fd < 0)
-			open_error(in->data);
-		else
-		{
-			dup2(fd, STDIN_FILENO);
-			close(fd);
-		}
-		in = in->next;
+		dup2(fd, STDOUT_FILENO);
+		close(fd);
 	}
 }
 
-void	redirect(t_content *cmd, char **main_arr, int pos)
+void	redirect(char **main_arr, int pos)
 {
-	int			tmp_pos;
-	t_content	*tmp_cmd;
+	int		index;
 
-	tmp_cmd = cmd;
-	tmp_pos = pos;
-	while (main_arr[tmp_pos] && main_arr[tmp_pos][0] && ft_strcmp(main_arr[\
-		tmp_pos], "|") != 0)
+	index = pos;
+	while (main_arr[index] && main_arr[index][0] && ft_strcmp(main_arr[\
+		index], "|") != 0)
 	{
-		if (ft_strcmp(main_arr[tmp_pos], ">") == 0)
-			redirect_out(tmp_cmd->output, 0);
-		if (ft_strcmp(main_arr[tmp_pos], ">>") == 0)
-			redirect_out(tmp_cmd->append, 1);
-		if (ft_strcmp(main_arr[tmp_pos], "<") == 0)
-			redirect_in(tmp_cmd->input);
-		if (ft_strcmp(main_arr[tmp_pos], "<<") == 0)
-			redirect_in(tmp_cmd->heredoc);
-		tmp_pos++;
+		if (ft_strcmp(main_arr[index], "<") == 0)
+			redirect_in(main_arr[index + 1], 0);
+		else if (ft_strcmp(main_arr[index], "<<") == 0)
+			redirect_in(main_arr[index + 1], 1);
+		else if (ft_strcmp(main_arr[index], ">") == 0)
+			redirect_out(main_arr[index + 1], 0);
+		else if (ft_strcmp(main_arr[index], ">>") == 0)
+			redirect_out(main_arr[index + 1], 1);
+		index++;
 	}
 }
