+++ src/signals.c
@@ -6,7 +6,7 @@
 /*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/16 16:08:59 by mortins-          #+#    #+#             */
-/*   Updated: 2023/10/30 18:38:23 by mortins-         ###   ########.fr       */
+/*   Updated: 2023/11/01 17:11:05 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,10 @@ void	signal_process_interrupt(int signum)
 void	signal_exit(t_minishell *ms)
 {
 	if (!(ms->str))
+	{
+		write(STDOUT_FILENO, "\n", 1);
 		free_ms(ms);
+	}
 }
 
 //	CTRL-C sends an "Interrupt" signal.
