+++ src/structs/content.c
@@ -6,7 +6,7 @@
 /*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/14 14:59:23 by mortins-          #+#    #+#             */
-/*   Updated: 2023/11/01 16:39:08 by mortins-         ###   ########.fr       */
+/*   Updated: 2023/11/01 17:12:20 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,8 @@
 
 void	init_heredoc(t_minishell *ms, char **main_arr)
 {
-	int	index;
+	int		index;
+	char	*buf;
 
 	index = 0;
 	if (!main_arr || !main_arr[0] || !main_arr[0][0])
@@ -23,40 +24,15 @@ void	init_heredoc(t_minishell *ms, char **main_arr)
 	{
 		if (ft_strcmp(main_arr[index], "<<") == 0)
 		{
+			buf = ft_strdup(main_arr[index + 1]);
 			free(main_arr[index + 1]);
-			main_arr[index + 1] = heredoc(ms, main_arr[index + 1], index);
+			main_arr[index + 1] = heredoc(ms, buf, index);
+			free(buf);
 			index += 2;
 		}
 		else
 			index++;
 	}
-}
-
-t_list	*hdoc_lst(t_minishell *ms, char **arr, int index)
-{
-	char	*file;
-	t_list	*node;
-	t_list	*hdoc;
-
-	hdoc = NULL;
-	while (arr[index] && ft_strcmp(arr[index], "|") != 0)
-	{
-		if (ft_strcmp(arr[index], "<<") == 0)
-		{
-			file = heredoc(ms, arr[index + 1], index);
-			node = ft_lstnew(file);
-			if (!node)
-			{
-				ft_lstclear(&hdoc, free);
-				return (NULL);
-			}
-			ft_lstadd_back(&hdoc, node);
-			index += 2;
-		}
-		else
-			index++;
-	}
-	return (hdoc);
 }
 
 t_list	*redir_lst(char **arr, int index, char *limiter)
